+++ src/utils/union_find.rs
@@ -1,18 +1,18 @@
 //! The union find algorithm.
 
 /// An element used by the union-find algorithm.
-#[deriving(Clone)]
+#[derive(Clone)]
 pub struct UnionFindSet {
     /// The parent of this union find element.
-    parent: uint,
+    parent: usize,
     /// The rank of this union find element.
-    rank:   uint
+    rank:   usize
 }
 
 impl UnionFindSet {
     /// Creates a new `UnionFindSet`.
     #[inline]
-    pub fn new(key: uint) -> UnionFindSet {
+    pub fn new(key: usize) -> UnionFindSet {
         UnionFindSet {
             parent: key,
             rank:   0
@@ -21,14 +21,14 @@ impl UnionFindSet {
 
     /// Reinitialize this set.
     #[inline]
-    pub fn reinit(&mut self, key: uint) {
+    pub fn reinit(&mut self, key: usize) {
         self.parent = key;
         self.rank   = 0;
     }
 }
 
 /// Performs the `find` part of the union-find algorithm.
-pub fn find(x: uint, sets: &mut [UnionFindSet]) -> uint {
+pub fn find(x: usize, sets: &mut [UnionFindSet]) -> usize {
     if sets[x].parent != x {
         sets[x].parent = find(sets[x].parent, sets);
     }
@@ -37,7 +37,7 @@ pub fn find(x: uint, sets: &mut [UnionFindSet]) -> uint {
 }
  
 /// Performs the `union` part of the union-find algorithm.
-pub fn union(x: uint, y: uint, sets: &mut [UnionFindSet]) {
+pub fn union(x: usize, y: usize, sets: &mut [UnionFindSet]) {
      let x_root = find(x, sets);
      let y_root = find(y, sets);
